requirements.txt
torch
torchvision
numpy
scipy
faiss-cpu
pyyaml
pytest
matplotlib
